Proceso se asignación de tarifas casi listo, se trabajo en la adminstracion de tarifas de viaje. Iniciando en nuevo branch Archivos ignorados agregados al repositorio. Coomit previo a merge para sincronizar con master

+++ src/main/java/mx/edu/utxj/pye/sgi/entity/ch/EventosAreas.java
@@ -17,7 +17,7 @@ import javax.xml.bind.annotation.XmlRootElement;
 
 /**
  *
- * @author UTXJ
+ * @author jonny
  */
 @Entity
 @Table(name = "eventos_areas", catalog = "capital_humano", schema = "")
@@ -42,7 +42,7 @@ public class EventosAreas implements Serializable {
         this.eventosAreasPK = eventosAreasPK;
     }
 
-    public EventosAreas(int evento, int areaOperativa) {
+    public EventosAreas(int evento, short areaOperativa) {
         this.eventosAreasPK = new EventosAreasPK(evento, areaOperativa);
     }
